+++ src/enums.ts
@@ -1,0 +1,28 @@
+
+export enum FindType {
+    Definition = 1,
+    Reference = 2,
+}
+
+export enum FindCommandType {
+    RegexFindDefinitionInCodeFiles,
+    RegexFindDefinitionInCurrentFile,
+    RegexFindReferencesInCurrentFile,
+    RegexFindReferencesInCodeFiles,
+    RegexFindPureReferencesInCodeFiles,
+    RegexFindReferencesInConfigFiles,
+    RegexFindReferencesInDocs,
+    RegexFindReferencesInAllProjectFiles,
+    RegexFindReferencesInAllSmallFiles,
+    RegexFindReferencesInCodeAndConfig,
+    FindPlainTextInCodeFiles,
+    FindPlainTextInConfigFiles,
+    FindPlainTextInDocFiles,
+    FindPlainTextInConfigAndConfigFiles,
+    FindPlainTextInAllProjectFiles,
+    FindPlainTextInAllSmallFiles,
+    SortProjectFilesBySize,
+    SortProjectFilesByTime,
+    SortAllFilesBySize,
+    SortAllFilesByTime,
+}